requirements.txt
# Core dependencies for the AI internship project
mysql-connector-python>=8.0.33
pdfplumber>=0.9.0
PyPDF2>=3.0.0

# Development and testing dependencies
pytest>=7.4.0
pytest-mock>=3.11.1

# Additional dependencies for future extensions
# These will be needed for weeks 3-4
requests>=2.31.0          # For API calls to ChatGPT
openai>=1.3.0            # OpenAI API client
pandas>=2.0.0            # Data manipulation
nltk>=3.8.1              # Natural language processing
scikit-learn>=1.3.0      # Machine learning utilities
joblib>=1.3.0            # Efficient serialization
python-dotenv>=1.0.0     # Environment variable management

# Optional dependencies for enhanced functionality
schedule>=1.2.0          # Task scheduling
click>=8.1.0             # CLI interface
colorlog>=6.7.0          # Colored logging
tqdm>=4.65.0             # Progress bars